Remove circular buffer files + fix warnings

+++ ProjetNorvege/flycameramanager.cpp
@@ -15,10 +15,10 @@ FlyCameraManager::FlyCameraManager()
 
     std::vector<CameraProperty> prop = std::vector<CameraProperty>();
     prop.push_back(CameraProperty(CameraManager::BRIGHTNESS, 0, 255, 0, true));
-    prop.push_back(CameraProperty(CameraManager::GAIN, 0, 12.041, 3, true));
+    prop.push_back(CameraProperty(CameraManager::GAIN, 0, (float) 12.041, 3, true));
     prop.push_back(CameraProperty(CameraManager::EXPOSURE, 0, 62, 0, true));
     prop.push_back(CameraProperty(CameraManager::GAMMA, 0, 1, 0, false));
-    prop.push_back(CameraProperty(CameraManager::SHUTTER, 0, 33.174, 3, true));
+    prop.push_back(CameraProperty(CameraManager::SHUTTER, 0, (float) 33.174, 3, true));
 
     for(int i=prop.size()-1; i>=0; i--){
         prop.at(i).setValue(0.0);
